requirements.txt
discord.py=2.0.1
beautifulsoup4=4.11.1
gspread=5.6.0
requests=2.28.1